# packages/backend/openapi.yml
openapi: 3.0.0
paths:
  /api/health:
    get:
      operationId: HealthController_get
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHealthDto'
      tags: &ref_0
        - Health
  /api/health/auth:
    get:
      operationId: HealthController_authCheck
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHealthDto'
      tags: *ref_0
  /api/todo:
    get:
      operationId: TodoController_get
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetTodoDto'
      tags: &ref_1
        - Todo
    post:
      operationId: TodoController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTodoDto'
      tags: *ref_1
    put:
      operationId: TodoController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTodoDto'
      tags: *ref_1
    delete:
      operationId: TodoController_delete
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTodoDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTodoDto'
      tags: *ref_1
info:
  title: Example
  description: The cats API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    GetHealthDto:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    GetTodoDto:
      type: object
      properties:
        id:
          type: string
        complete:
          type: boolean
        title:
          type: string
        comment:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - complete
        - title
        - comment
        - createdAt
        - updatedAt
    CreateTodoDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        comment:
          type: string
          minLength: 1
      required:
        - title
        - comment
    UpdateTodoDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 1
        comment:
          type: string
          minLength: 1
        complete:
          type: boolean
      required:
        - id
        - title
        - comment
        - complete
    DeleteTodoDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id